requirements.txt
absl-py
anyio
axial_positional_embedding
certifi
click
CoLT5-attention
contourpy
cycler
easy-torch
easydict
einops
exceptiongroup
filelock
fonttools
fsspec
grpcio
h11
hf-xet
httpcore
httpx
huggingface_hub
hyper-connections
idna
ImageIO
Jinja2
joblib
kiwisolver
lazy_loader
local-attention
Markdown
MarkupSafe
matplotlib
mpmath
networkx
numpy
opencv-contrib-python-headless
packaging
pandas
pillow
product_key_memory
protobuf
pyparsing
python-dateutil
pytz
PyYAML
reformer-pytorch
safetensors
scikit-image
scikit-learn
scipy
seaborn
setproctitle
shellingham
six
sympy
tensorboard
tensorboard-data-server
threadpoolctl
tifffile
timm
tqdm
typer-slim
typing_extensions
tzdata
Werkzeug